Added images and text
Also styled two pages for wide screen as well ass mobile screen. Still have to  fix a few things on second page and style the third page. Need some help with the navigation bar. Want Option 2  and Option 3 to fit on one line, not create two. Looks like the space is too small, don't know how to make it wider for the whole text to fit.